api_service/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
aio-pika==9.5.7
pydantic==2.11.7
tortoise-orm==0.25.1